PatitasFelices - Backend:   - None
- Frontend:
  - Correccion de importaciones
  - Comente archivos que no son utilizados, proximamente hay que borrarlos
